showing pokemon list + base jetpack navigation

+++ core/src/main/kotlin/com/badront/pokedex/core/model/Page.kt
@@ -1,6 +1,6 @@
 package com.badront.pokedex.core.model
 
 class Page<T>(
-    val total: Long,
+    val total: Int,
     val items: List<T>
 )
