+++ core/src/main/kotlin/com/badront/pokedex/core/model/PageInfo.kt
@@ -2,5 +2,5 @@ package com.badront.pokedex.core.model
 
 class PageInfo(
     val limit: Int,
-    val offset: Long
+    val offset: Int
 )
